move to app/components/featureFlags
