+++ static/app/components/featureFlags/hooks/useOrganizationFlagLog.tsx
@@ -1,6 +1,6 @@
+import type {RawFlagData} from 'sentry/components/featureFlags/utils';
 import type {Organization} from 'sentry/types/organization';
 import {useApiQuery} from 'sentry/utils/queryClient';
-import type {RawFlagData} from 'sentry/views/issueDetails/streamline/featureFlagUtils';
 
 export function useOrganizationFlagLog({
   organization,
